requirements.txt
# FastAPI and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1

# Database
oracledb==2.0.1
sqlalchemy==2.0.25

# Celery and Redis
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Logging and Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
